prettier & nicer tester... and some changes in MINISHELL

+++ tests/tester.sh
@@ -27,7 +27,8 @@ main() {
 	elif [[ $1 == "a" ]] ; then
 		test_mandatory
 		test_bonus
-		# test_from_file "cmds/test.sh"
+	elif [[ $1 == "-f" ]] ; then
+		[[ -f $2 ]] && test_from_file $2
 	else
 		echo "usage: bash test.sh [m,b,a]"
 	fi
@@ -82,10 +83,12 @@ test_from_file() {
 	IFS=''
 	i=1
 	end_of_file=0
+	line_count=0
 	while [[ $end_of_file == 0 ]] ;
 	do
 		read -r line
 		end_of_file=$?
+		((line_count++))
 		if [[ $line == \#* ]] || [[ $line == "" ]] ; then
 			# if [[ $line == "###"[[:blank:]]*[[:blank:]]"###" ]] ; then
 			# 	echo -e "\033[0;33m$line\033[m"
@@ -95,11 +98,13 @@ test_from_file() {
 			continue
 		else
 			printf "\033[0;35m%-4s\033[m" "$i:"
+			tmp_line_count=$line_count
 			while [[ $end_of_file == 0 ]] && [[ $line != \#* ]] && [[ $line != "" ]] ;
 			do
 				INPUT+="$line$NL"
 				read -r line
 				end_of_file=$?
+				((line_count++))
 			done
 			$MINISHELL_PATH/$EXECUTABLE <<< "$INPUT" 2>tmp_err_minishell >tmp_out_minishell
 			exit_minishell=$?
@@ -108,7 +113,7 @@ test_from_file() {
 			echo -ne "\033[0;34mSTD_OUT:\033[m "
 			if ! diff -q tmp_out_minishell tmp_out_bash >/dev/null ;
 			then
-				echo -ne "\033[0;31mKO" [ \"${INPUT:0:15}\" ... ] "\033[m  " | tr '\n' ' '
+				echo -ne "\033[0;31mKO\033[m  " | tr '\n' ' '
 				((TEST_KO_OUT++))
 			else
 				echo -ne "\033[0;32mOK\033[m  "
@@ -116,7 +121,7 @@ test_from_file() {
 			echo -ne "\033[0;36mSTD_ERR:\033[m "
 			if [[ -s tmp_err_minishell && ! -s tmp_err_bash ]] || [[ ! -s tmp_err_minishell && -s tmp_err_bash ]] ;
 			then
-				echo -ne "\033[0;31mKO" "[ \"${INPUT:0:15}\" ... ]" "\033[m  " |  tr '\n' ' '
+				echo -ne "\033[0;31mKO\033[m  " |  tr '\n' ' '
 				((TEST_KO_ERR++))
 			else
 				echo -ne "\033[0;32mOK\033[m  "
@@ -124,7 +129,7 @@ test_from_file() {
 			echo -ne "\033[0;36mEXIT_CODE:\033[m "
 			if [[ $exit_minishell != $exit_bash ]] ;
 			then
-				echo -ne "\033[0;31mKO [ minishell($exit_minishell)  bash($exit_bash) ]" [ \"${INPUT:0:15}\" ... ] "\033[m  " | tr '\n' ' '
+				echo -ne "\033[0;31mKO [ minishell($exit_minishell)  bash($exit_bash) ]\033[m  " | tr '\n' ' '
 				((TEST_KO_EXIT++))
 			else
 				echo -ne "\033[0;32mOK\033[m  "
@@ -132,7 +137,7 @@ test_from_file() {
 			INPUT=""
 			((i++))
 			((TEST_COUNT++))
-			echo ""
+			echo -e "\033[0;90m$1:$tmp_line_count\033[m  "
 		fi
 	done < "$1"
 }
@@ -141,4 +146,4 @@ test_from_file() {
 main "$@"
 
 # Clean all tmp files
-rm -f tmp_*
+[[ $1 != "-f" ]] && rm -f tmp_*
